Allow tests to read spec files from `spec_lib/` directory
This removes the need to duplicate a spec file under `tests/` to test it.
If a spec file is present under `tests/`, it's still preferred over the one
under `spec_lib/`

This commit also moves the `spec_lib/` files under a subdirectory for the arch,
to match the test runner's expectation that testable `spec_lib/` files
are under a subdirectory with the same name as the test directory. This isn't
strictly necessary since the test runner already requires that sub-test names
are also unique across all test directories, but it's useful for RISC-V
specifically since there will likely be more `rv*.isa` later for extensions.